feat(pci-project): test to rework
ref: #MANAGER-18604
