+++ packages/manager/apps/pci-project/src/tests/index.ts
@@ -1,0 +1,15 @@
+// Tests utilities - PCI Project App
+// Main entry point for all test utilities
+
+// Re-export everything from helpers (includes mocks)
+export * from './helpers';
+
+// Re-export wrappers for convenience
+export { shellContext } from './helpers/lightweight-wrappers';
+
+export {
+  createOptimalWrapper,
+  createMinimalWrapper,
+  createNavigationWrapper,
+  createShellWrapper,
+} from './helpers/lightweight-wrappers';
