+++ packages/manager/apps/pci-project/src/tests/mocks/setup.ts
@@ -6,7 +6,7 @@ import {
   mockHooks,
   mockConstants,
   mockI18n,
-} from './index';
+} from '../helpers';
 
 // Setup all mocks globally
 export const setupAllMocks = () => {
